selectic: Hide optgroup name for few items.
When there are less items then the page size, there is no reason to add
the line to display the group name.

So now it is displayed only when there is a scroll bar.

The number of items displayed in a page is now handled by Store to have a
common value through all components.

Change-Id: I7b5703aa609e8af7a105a5b98468cc5945703286