mainInput: fix double clear icon.
When Selectic is in multiple mode and only one item is selected then a
double icon to remove it was displayed.

This is is because the first icon is related to the item itself and the
second to clear all items (so the lone selected item).

Now, if there is only one selected item the icon to clear the whole
selection is not displayed anymore.

Change-Id: Ie9f8ca3772a96afd11bccb2f7a4aff02f7a88883
Closes: #67789