Selectic: fix clicking outside body.
When the Selectic component is open, if user click out the body element
(if page is short), Selectic stayed open.

This was because event listener was added to document.body and body stay
as small as possible.
So user can click outside <body> and listener was not triggered.

The listener is now added to document, like this it works when user clicks
on body and outside body (but inside the page window).

Change-Id: Id603e5a21c773e5790d0bd7162428cff866bcd4b
Closes: #68084