Fix list position.
With Selectic located in CSS grid with scroll, the position of the list
was wrongly computed.

It was because the scroll container was not in the offset parents. And so
it was neither recomputed (there was no listener) nor correctly computed
(surprisingly the offsetTop doesn't include the scroll inside a CSS grid
layer).

Now a bounding box is computed around the element. This technique is now
quite well supported (all Edge versions) to be used.

Change-Id: I0b3393ffffcffafad86397b26bc5871c850e0486
Closes: #72148