Rename inner properties.
Properties named offsetTop was related to the bottom of the main element.
It was quite confusing to read top and think bottom. So this property and
similar ones are renamed to elementBottom referring to the main element.

It is now at the usage that we should think about which one we want to
use.

Change-Id: I3d8cbfd2034e45d8868ae7287c99e7ad691954a2