Move panel horizontal position depending on window viewport
If Selectic is position at right of the page and is quite small (so the
panel is greater than the main element), it could create an horizontal
scroll as the panel extends outside the current view port.

Now it checks where the panel will be displayed and choose the best
position.
First, if it can be extended on the right.
Then, if it can be extended on the left.
Finally don't expand the panel (even if the size is quite small).

Change-Id: I16e1662c792b62bdbf80078d421e60d6141ba1fc