Add "open" and "close" events
The events "open" and "close" have been added.

This can be useful when some action are needed when user interacts with
Selectic.
These are quite equivalent to "focus" and "blur".

The reference to the selectic instance has been added to arguments so
methods can be used more easily without using the "ref" attributes.

Documentation has been updated (and also for other events)

Change-Id: Icdd64b58c1184c894e616731fb8f85cd0f3c90c2