Embed CSS.
CSS is know embedded in the library.
So it is not necessary to load any CSS files or to rewrite the full
styles.

Thanks to CSS variable, the default values can be changed easily in any
project.

As the CSS is injected in the header, it should be checked how it should
work with CSP (and change it or document it if it is necessary).

Change-Id: Ie46450affdf2922492b2e456e3c857880ca8db96