fix auto-disabled with invalid value.
Selectic has a feature to disable the component when there are no items or
only one and this value is selected.

The problem is that the current selected value may be an invalid value
(coming from previous options which have been changed).
So the component is disabled and it is not possible to select the correct
one.

So now if the value is invalid, the component stay enabled.
If the strictValue is active (and autoSelect which is true by default),
the value is changed to the first valid value and then the component can
be disabled.

Many tests have also been fixed (due to miswriting or which forgetting the
autoselect behavior (which was not working correctly in some cases)).

Change-Id: Ic8af46753e379d6c53c66d636ceb1e2672d5ca47