Update dependencies.
updates VTYX to 3.0, which means that it is no more needed to have a peer
dependencies on JQuery and Bootstrap!

Tape cannot be update to 5.X because deepEqual has changed its behavior
and it doesn't pass tests. Although "is-equal/why" didn't show any
difference, the tests fails. I am guessing that it is due to Vue which
adds some hidden properties.
So until then Tape stay stuck in 4.X.

Change-Id: I32ceff4010a1ab82be9af0a2ce59a5815c0fab08