Fix displaying icon after caret.
To avoid displaying text and icon under the caret, the bootstrap
equivalent style has been added. However this one is dependent on
Selectic size choices.

Title have been added on single selected item and placeholder in case
they are truncated with ellipsis. A better solution (to display the
title only when the text is truncated) should be found and developed.

Change-Id: Ib1608d13c7c9cee3c47fffc39f6b1a65c8d76cc2