Use unref.
ref and computed properties, can be read by adding a .value, but in Vue
context the property is unwrapped and so can be read without the .value.

For some methods, the attributes can be read in different condition and
so we had to write so odd line to check if the property is a ref value
or not.

But Vue includes a method which simplify this a lot (and Typescript
understand much better what we are doing): unref

Change-Id: Iff17a41e7c80d7caffb2c3674e679ba50929b228