Fix issue when clicking outside when focus is lost.
When focus is lost, the panel should be closed (but it depends on
options), but if we click outside before the events are removed, it
generates an error because we do not have a reference to the panel
element.

So now we check if the panel element is in reference.

Change-Id: I3e0bc7fad12d8b9e642d1d50aaa92bbfd127cb32