Fix again issue with sticky disabled.
Now the issue comes from the !! which convert to true when the value is
a Vue proxy and so the component is disabled (even if the proxy value
should be false).

Because the internal attributes this.props has been changed with
reactive, there is no need to use props with unref.

Change-Id: I863f00652cacc3aa048b65fedd4ff1ecd67ec3ef