Fix props modification.
When parent gives a new options, this props was given to the store.
Only the array was duplicated, but not inner values.
So when stores modified inner options, it modifies the parent original
object.

In some case these modifications introduces infinite recursive loop.

To avoid such issue, the options are now deeply clone.

This function (and some other helpers) have been added to the tools
file.

Change-Id: I021a0b709f7bc834154ed15c7c53d848957bc49e