Add exclusive option.
It is now possible to add a "exclusive" property on option items.

In multiple mode, an option which has this property cannot be selected
with other options. If there are already selected options then they will
be deselected when an exclusive option is selected. An exclusive option
will be automatically deselected when another option is selected.

Change-Id: I34ffe14b410e5b5a983986192f0783ec25aa7683