Disable "select all" when data is missing.
In multiple mode, there is a "select all" action. It allows user to
select all items at once.
But in dynamic mode, items can be fetched partially. In such case, the
revert selection is the solution to select all items that have not been
fetched yet.

So if this option has been disabled, the "select all" has no reason to
be enabled.

An option ('forceSelectAll') has been added if the previous behavior is
the expected one.

In the future, we may add options, such as to force fetch all missing
items when using this action.

Change-Id: Id719e2fb17577c9d5a2730ca23126d68027873d4