Improve list position.
When there is not enough space, the list reduce its height to avoid
hiding some parts.
But sometimes, it reduces while there is enough space above the input.

This was due to this.listHeight which remembers the size in the reduce
height.

To fix this behavior, a getter "isFullyEstimated" has been added to know
if we can trust the listHeight value or not.
If we cannot trust this value then we compute the best position on an
estimated list height (with 10 options displayed).

Change-Id: I972b3e9a30346a72359aee24d29941b5548d2c1c