improve list UI when displaying at top.
The list was displayed with a box shadow, but the shadow was always
displayed at bottom which overlapped the input.

Now, a class is added to the list, so it is possible to customize the
style depending on its position.

By default, the shadow is now at top when the list is at top.

Change-Id: If7db4b2d9538b936a4b9a305bbb86430bfb98a7e