[Fix] when rendered out of DOM
When Selectic is created and rendered out of DOM, it cannot compute the
correct size for the list.
But when it is attached to the DOM, the computation was not done again.

Now, if Selectic is not in DOM when rendered, an observer is added which
re-trigger the size computation when Selectic is attached to DOM.

Change-Id: I7214e39cfddecf0ad7308ae1409c14e00dd9671d