Display title on selected items
In multiple mode, on selected items we display a title to show the full
text when there is not enough space to display them.

But if the item description contains a title, it would be more logical
to display it instead of the text.

Also displays the title first on single selection.

Change-Id: Ic27ac10f041fb03ca33e34f47933f9eeec8f4bd2