Fix auto-disabled
When switching options, Selectic should be disabled if there is only one
option. But if there was a selected value and this is not the option
value, we should not disable Selectic right now.
It should be disabled when user does not have any more choice.

For this purpose, we should check the auto-disabled when changing value.

Change-Id: I69f6dab0cafdfd7ec9c4611eadbdd0bff3f5f0cf