Fix error message display
When looping on similar values (in dynamic mode), there is a protection
to avoid looping infinitely. Unfortunately in such case the code to
check if the error message has to be displayed read a computed variable
but failed to read the "value".
This is now done with the unref method (which works if the value is a
simple variable or a ref variable).

Change-Id: I8b0c83e0eb614c7ccb449107b964ff02f22482db