Fix dynamic mode with latency
When requests have latencies there are high chance to repeat several
requests.
Unfortunately there is also high chance to request for something less
than a previous request.
So the cache system contains already the data and declare that the
request is bad (protection against missing total).

To avoid such issues, only the latest request is handled and the cache
is no more built with previous requests.
A debounce is done when several request should be sent at the same time.

The filteredOptions was also reset too often, now a dedicated method has
been created and is called only when needed.

Change-Id: If3303fe34c833e61912eb6a4fbdabefab249d7e4