Prevent issue when $refs is accessed.
It happens that some methods are called while $refs is undefined or
their properties are not set yet.

It should not happen but can appear if the methods are called too soon
(they are called before elements are created (in render) or too late
(the method is called after the unmount).

This commit only adds protection to avoid failing in such cases. But
these cases should be avoided as much as possible.

Change-Id: Ib119b704a83f56f45e66b8de545a3c8d7b65e6fe