Fix autodisabled with selected disabled items.
In multiple mode,
When a disabled item is already selected, we should not disable Selectic
if the only enabled item is selected because it is still valid to
unselect it (we still have a value which is the disabled item).

A special case was also be treated which is when an exclusive item is
disabled and selected. Then Selectic is disabled because user cannot do
anything (we should modify the value property).

Change-Id: I610dc2cc427cd210a540d3a4d8bdfa050034a906