Fix manipulation with selected disabled items.
In multiple mode,
When a disabled item is already selected, it should not be removed from
the selection (it is considered as a mandatory item).

Consequently, the "clear" action should not remove them, but also when
clicking on "exclusive" item (because this exclusive item do not share
their space with others).

In "simple" mode,
A disabled item should not avoid changing selection. However we cannot
unselect it directly (because it is disabled) but we can still clear the
selection.

Change-Id: Ib0bf75e6ef18a911d4522d3740f0aa12c417377a