Fix keydown event handling
What:
In some cases, when clicking on the enter button on the keyboard in a
text area and there is a selectic element in the page, the line break
does not occur.

Why:
removeEventListener is not working properly so selectic blocks the
default behavior of the "Enter", "ArrowUp", and "ArrowDown" buttons
because they are used for manipulating the select with the keyboard.

How:
Instead of passing a computed into removeEventListener which could
return a different function reference each time (thus breaking the
removeEventListener), use a method.

Change-Id: I0d51ca2184c0f20cedee43c908df818b23880a1f