Hide icon on hover for single select elements
On hover, on single select list (ie without the "multiple" option), the
dot icon appear as a 'placeholder'. While this behavior is useful for
multiple select list, it could be confusing for single select list and
give the idea that the user can select multiple element.

This patch hide the 'placehorder' dot icon that appear on hover, on
un-selected elements.
Note that the check icon for multiple select list will still appear on
hover, as before.

Change-Id: I038af3e7484f66bf354ca6e0471e4a0ec3680fa5